requirements.txt
mysql-connector-python==8.2.0
PyYAML==6.0.1
setuptools==80.9.0
pytest==8.3.4
